fix negative lookahead over choice: -(foo,bar)
Previously this rule syntax caused an error because Julia parsed it as a
call to the :- function with multiple arguments, rather than as the
unary negation that PEG expected. To get the unary negation reading you
would have had to double the parens: -((foo,bar)). This change adds
another method to the to_rule function that translates the
multiple-argument version to the single-tuple-argument version
automatically, so that you no longer need to double the parens.